+++ src/onegov.activity/onegov/activity/collections/publication_request.py
@@ -1,0 +1,9 @@
+from onegov.activity.models.publication_request import PublicationRequest
+from onegov.core.collection import GenericCollection
+
+
+class PublicationRequestCollection(GenericCollection):
+
+    @property
+    def model_class(self):
+        return PublicationRequest